Move the code from master/main to master.py
